Create a map of RAWS stations

Generates a map of RAWS stations given station metadata. This station metadata can be generated by wrcc_createMeta() or cefa_createMeta(). Parameters such as radius, stroke, color, opacity, fillColor fillOpacity, and weight can be passed in to adjust features of the map.

meta_leaflet(
  meta = NULL,
  radius = 3.5,
  stroke = TRUE,
  color = "blue",
  opacity = 0.5,
  fillColor = "blue",
  fillOpacity = 0.5,
  weight = 1
)

Arguments

meta

Station metadata created by wrcc_createMeta() or cefa_createMeta().

radius

Circle radius.

stroke

Logical specifying whether to draw stroke along the path.

color

Stroke color.

opacity

Stroke opacity.

fillColor

Fill color.

fillOpacity

Fill opacity.

weight

Stroke width in pixels.

Value

Leaflet map of stations.

References

RAWS USA Climate Archive

See also

Examples

# \donttest{
library(RAWSmet)

wa_meta <- example_wrcc_meta
meta_leaflet(wa_meta)

# }