Add wind barbs to a map for multiple RAWS stations

Add a multi-sided polygon to a plot.

rawsList_addWindBarbs(
  rawsList = NULL,
  observationTime = NULL,
  skipMissing = FALSE,
  type = "winds",
  circleSize = 1,
  circleFill = "transparent",
  lineCol = 1,
  extraBarbLength = 0,
  barbSize = 1,
  ...
)

Arguments

rawsList
a list of raws_timeseries objects

observationTime
Time of observation in question. Only accepts POSIXct objects.

skipMissing
Logical flag to skip stations with data missing at the requested time.

type
Which type of wind to plot. Currently only accepts 'winds' or 'gusts'

circleSize
size of the circle

circleFill
circle fill color

lineCol
line color (currently not supported)

extraBarbLength
add length to barbs

barbSize
size of the barb

...
additional arguments to be passed to lines

References

https://commons.wikimedia.org/wiki/Wind_speed

Examples

if (FALSE) {
library(maps)
library(RAWSmet)

setRawsDataDir("~/Data/RAWS")

maps::map('state', "washington")

# Noon on 09-01-2017
observationTime <- MazamaCoreUtils::parseDatetime(2017090112, timezone = "America/Los_Angeles")

rawsList_addWindBarbs(rawsList = example_fw13List, observationTime = observationTime)
}