Convert a list of raws_timeseries objects to a single rawsDF object

Converts a list of raws_timeseries objects (raws_list) to a single, tidy dataframe each containing all variables from each raws_timeseries object's data dataframe along with the following values from their meta dataframes:

nwsID - the nwsID of the station

wrccID - the wrccID of the station

siteName - the name of the station

longitude - longitude coordinate of the station

latitude - latitude coordinate of the station

timezone - the timezone the station is in

elevation - elevation of the station

This version of the RAWS data is known as a rawsDF object.

Replicating this set of variables for every record greatly inflates the size of the data but also makes it much more useful when working with the dplyr and ggplot2 packages.

Multiple rawsDF objects can be combined with dplyr::bind_rows() and used to create multi-station plots.

rawsList_toRawsDF(rawsList = NULL)

Arguments

rawsList
List of raws_timeseries objects.

Value

A single tidy dataframe containing data and metadata for each raws_timeseries object in a raws_list.

See also

Examples

library(RAWSmet)

rawsDF<- example_fw13List %>% rawsList_toRawsDF()

head(rawsDF)

#> # A tibble: 6 x 19
#>   datetime            temperature humidity windSpeed windDirection maxGustSpeed
#>   <dttm>                    <dbl>    <dbl>     <dbl>         <dbl>        <dbl>
#> 1 2017-09-01 07:00:00        35       46.5      2.68            58         5.81
#> 2 2017-09-01 08:00:00        35.6     46        2.68            72         4.47
#> 3 2017-09-01 09:00:00        35.6     46        2.68            61         4.92
#> 4 2017-09-01 10:00:00        34.4     46        3.13            51         4.47
#> 5 2017-09-01 11:00:00        30       46        1.79            67         3.58
#> 6 2017-09-01 12:00:00        22.8     46        1.34           270         2.24
#> # … with 13 more variables: maxGustDirection <dbl>, precipitation <dbl>,
#> #   solarRadiation <dbl>, fuelMoisture <dbl>, fuelTemperature <chr>,
#> #   monitorType <chr>, nwsID <chr>, wrccID <chr>, siteName <chr>,
#> #   longitude <dbl>, latitude <dbl>, timezone <chr>, elevation <dbl>