Extract dataframes from raws_timeseries objects

These functions are convenient wrappers for extracting the dataframes that comprise a raws_timeseries object. These functions are designed to be useful when manipulating data in a pipeline chain using %>%.

raws_extractData(rawsObject = NULL, forOpenair = FALSE)

raws_extractMeta(rawsObject = NULL)

Arguments

rawsObject
raws_timeseries object to extract dataframe from

forOpenair
Logical instructions to clean data for use with openair

Value

A dataframe from the given raws_timeseries object

Examples

# \donttest{
library(RAWSmet)

data <-
  example_fw13SaddleMountain %>%
  raws_extractData(forOpenair = TRUE)

# openair::timePlot(
#  data,
#  pollutant = "temperature"
# )

example_fw13SaddleMountain %>%
  raws_extractMeta() %>%
  head()

#> # A tibble: 1 x 11
#>   nwsID wrccID nessID siteName longitude latitude timezone elevation countryCode
#>   <chr> <chr>  <chr>  <chr>        <dbl>    <dbl> <chr>        <dbl> <chr>      
#> 1 4527… NA     NA     SADDLE …     -120.     46.7 America…       650 US         
#> # … with 2 more variables: stateCode <chr>, agency <chr>

# }