Date filtering for raws_timeseries objects

Subsets a raws_timeseries object by date. This function always filters to day-boundaries.

Dates can be anything that is understood by lubridate::ymd() including either of the following recommended formats:

"YYYYmmdd"

"YYYY-mm-dd"

raws_filterDate(
  rawsObject = NULL,
  startdate = NULL,
  enddate = NULL,
  days = NULL,
  weeks = NULL,
  timezone = NULL
)

Arguments

rawsObject
raws_timeseries object.

startdate
Desired start datetime (ISO 8601).

enddate
Desired end datetime (ISO 8601).

days
Number of days to include in the filterDate interval.

weeks
Number of weeks to include in the filterDate interval.

timezone
Olson timezone used to interpret dates.

Value

A subset of the given raws_timeseries object.

Note

The returned data will run from the beginning of startdate until the beginning of enddate -- i.e. no values associated with enddate will be returned.

See also

Examples

# \donttest{
library(RAWSmet)

rawsObject <- example_fw13SaddleMountain

data201708 <- rawsObject %>% raws_filterDate(20170801, 20170901, timezone = "America/Los_Angeles")
head(data201708$data)

#> # A tibble: 6 x 12
#>   datetime            temperature humidity windSpeed windDirection maxGustSpeed
#>   <dttm>                    <dbl>    <dbl>     <dbl>         <dbl>        <dbl>
#> 1 2017-08-01 07:00:00        38.3     36.5      2.68           288         4.47
#> 2 2017-08-01 08:00:00        38.9     36.5      2.68           240         5.36
#> 3 2017-08-01 09:00:00        39.4     37        1.34           259         4.47
#> 4 2017-08-01 10:00:00        38.3     37        1.79           301         3.13
#> 5 2017-08-01 11:00:00        35       37        1.34           292         2.24
#> 6 2017-08-01 12:00:00        26.1     37        1.79           283         2.68
#> # … with 6 more variables: maxGustDirection <dbl>, precipitation <dbl>,
#> #   solarRadiation <dbl>, fuelMoisture <dbl>, fuelTemperature <chr>,
#> #   monitorType <chr>

# }