Identify RAWS station type

Examine the column names of the incoming character vector to identify different types of monitor data provided by RAWS.

The return is a list includes everything needed to identify and parse the raw data using readr::read_tsv():

monitorType -- identification string

rawNames -- column names from the data (including special characters)

columnNames -- assigned column names (special characters repaced with '.')

columnTypes -- column type string for use with readr::read_csv()

The monitorType will be one of:

"WRCC_TYPE1" -- ???

"WRCC_TYPE2" -- ???

"WRCC_TYPE3" -- ???

"UNKOWN" -- ???

wrcc_identifyMonitorType(fileString = NULL)

Arguments

fileString
character string containing RAWS data

Value

List including monitorType, rawNames, columnNames and columnTypes.

References

RAWS USA Climate Archive

Examples

if (FALSE) {
library(RAWSmet)

fileString <- wrcc_downloadData(wrccID = 'WENU')
monitorTypeList <- wrcc_identifyMonitorType(fileString)
}