Load WRCC RAWS timeseries object from a local directory

Loads WRCC station metadata and data from the rawsDataDir. If the data is not in this directory, this will download and save the data.

wrcc_load(
  wrccID = NULL,
  meta = NULL,
  year = NULL,
  forceDownload = FALSE,
  baseUrl = "https://wrcc.dri.edu/cgi-bin/wea_list2.pl",
  verbose = TRUE
)

Arguments

wrccID
RAWS station identifier (will be upcased)

meta
Tibble of FW13 station metadata.

year
Year to access station data for.

forceDownload
Logical flag stating whether or not to download and override existing data.

baseUrl
Base URL for data queries.

verbose
Logical flag controlling detailed progress statements.

Value

Timeseries object with 'meta' and 'data'.

References

Program for Climate, Ecosystem and Fire Applications

See also

Examples

if (FALSE) {
library(RAWSmet)

setRawsDataDir("~/Data/RAWS/")

stationMeta <- wrcc_loadMeta(stateCode = "WA")
rawsObject <- wrcc_load(wrccID = "waWENU", meta = stationMeta, year = 2020)

dplyr::glimpse(rawsObject)
}